Extract AppDelegate Tests
